requirements.txt
# 智能理财助手系统依赖

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# HTTP请求
requests>=2.31.0
aiohttp>=3.9.0

# AI引擎
openai>=1.0.0

# 定时任务
apscheduler>=3.10.0

# 数据解析
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 日期处理
python-dateutil>=2.8.0

# 环境变量管理
python-dotenv>=1.0.0

# 通知推送
smtplib  # 内置库

# 日志
loguru>=0.7.0

# 数据验证
pydantic>=2.0.0
